Add files via upload
Implemented the solution for the transportation problem using linear programming techniques. Added functions to formulate and solve the problem with PuLP and pandas. Also included a visualisation function using matplotlib and seaborn to display the results. The notebook includes detailed steps for problem definition, model formulation, and Python implementation, along with comprehensive explanations and example data.